Make all pipes polymorphic in the return value.
Compression and decompression pipes work like filter, so they should be
polymorphic in the return value to make composition easier.

Since anything after a `forP` will only be executed if the upstream pipe
has terminated, a call to `discard` there will not alter the pipe's
behavior, but will have the effect of making its return type
polymorphic, as desired.